+++ build-tests/tests/fail/net_without_tcp_missing_tcp.rs
@@ -1,0 +1,4 @@
+use build_tests::tokio_net::tcp;
+
+fn main() {}
+