+++ build-tests/tests/fail/tokio_without_net_missing_net.rs
@@ -1,0 +1,3 @@
+use build_tests::tokio::net;
+
+fn main() {}