+++ tokio-process/src/unix/mod.rs
@@ -11,7 +11,7 @@
 //!
 //! Our best approximation here is to check *all spawned processes* for all
 //! SIGCHLD signals received. To do that we create a `Signal`, implemented in
-//! the `tokio-signal` crate, which is a stream over signals being received.
+//! the `tokio-net` crate, which is a stream over signals being received.
 //!
 //! Later when we poll the process's exit status we simply check to see if a
 //! SIGCHLD has happened since we last checked, and while that returns "yes" we
@@ -29,9 +29,8 @@ use self::reap::Reaper;
 use super::SpawnedChild;
 use crate::kill::Kill;
 
-use tokio_net::driver::Handle;
+use tokio_net::signal::unix::{signal, Signal, SignalKind};
 use tokio_net::util::PollEvented;
-use tokio_signal::unix::{Signal, SignalKind};
 
 use mio::event::Evented;
 use mio::unix::{EventedFd, UnixReady};
@@ -96,13 +95,13 @@ impl fmt::Debug for Child {
     }
 }
 
-pub(crate) fn spawn_child(cmd: &mut process::Command, handle: &Handle) -> io::Result<SpawnedChild> {
+pub(crate) fn spawn_child(cmd: &mut process::Command) -> io::Result<SpawnedChild> {
     let mut child = cmd.spawn()?;
-    let stdin = stdio(child.stdin.take(), handle)?;
-    let stdout = stdio(child.stdout.take(), handle)?;
-    let stderr = stdio(child.stderr.take(), handle)?;
+    let stdin = stdio(child.stdin.take())?;
+    let stdout = stdio(child.stdout.take())?;
+    let stderr = stdio(child.stderr.take())?;
 
-    let signal = Signal::with_handle(SignalKind::sigchld(), handle)?;
+    let signal = signal(SignalKind::child())?;
 
     Ok(SpawnedChild {
         child: Child {
@@ -203,7 +202,7 @@ pub(crate) type ChildStdin = PollEvented<Fd<process::ChildStdin>>;
 pub(crate) type ChildStdout = PollEvented<Fd<process::ChildStdout>>;
 pub(crate) type ChildStderr = PollEvented<Fd<process::ChildStderr>>;
 
-fn stdio<T>(option: Option<T>, handle: &Handle) -> io::Result<Option<PollEvented<Fd<T>>>>
+fn stdio<T>(option: Option<T>) -> io::Result<Option<PollEvented<Fd<T>>>>
 where
     T: AsRawFd,
 {
@@ -224,6 +223,5 @@ where
             return Err(io::Error::last_os_error());
         }
     }
-    let io = PollEvented::new_with_handle(Fd { inner: io }, handle)?;
-    Ok(Some(io))
+    Ok(Some(PollEvented::new(Fd { inner: io })))
 }