+++ tokio-timer/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/tokio-timer/0.3.0-alpha.1")]
+#![doc(html_root_url = "https://docs.rs/tokio-timer/0.3.0-alpha.2")]
 #![warn(
     missing_debug_implementations,
     missing_docs,
@@ -31,10 +31,10 @@
 //!
 //! [`Delay`]: struct.Delay.html
 //! [`DelayQueue`]: struct.DelayQueue.html
-//! [`Throttle`]: throttle/struct.Throttle.html
+//! [`Throttle`]: throttle::Throttle
 //! [`Timeout`]: struct.Timeout.html
 //! [`Interval`]: struct.Interval.html
-//! [`Timer`]: timer/struct.Timer.html
+//! [`Timer`]: timer::Timer
 
 pub mod clock;
 pub mod delay_queue;