+++ ui-tests/src/lib.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "tokio-executor")]
-pub use tokio_executor;
-
-#[cfg(feature = "tokio")]
-pub use tokio;