+++ tokio-test/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0-alpha.2 (August 17, 2019)
+
+### Changed
+- Update `futures` dependency to 0.3.0-alpha.18.
+
 # 0.2.0-alpha.1 (August 8, 2019)
 
 - Initial release